+++ mmodel/model.py
@@ -1,10 +1,11 @@
 import inspect
-from mmodel.utility import parse_input
+from mmodel.utility import parse_input, is_node_attr_defined, is_edge_attr_defined
+import networkx as nx
 
 
 class Model:
-    """Create the final executor
-    
+    """Create model executable
+
     :param object model_graph: ModelGraph instance
     :param class handler: Handler class that handles model execution
     :param dict handler_args: additional handler argument that is not model_graph
@@ -19,6 +20,8 @@ class Model:
     def __init__(
         self, model_graph, handler, handler_args={}, modifiers=[], add_returns=[]
     ):
+
+        assert self.is_graph_valid(model_graph)
 
         self.__name__ = f"{model_graph.name} model"
         self.model_graph = model_graph.copy()
@@ -55,3 +58,37 @@ class Model:
                 f"{self.model_graph.graph.get('doc', '')}",
             ]
         )
+
+    @staticmethod
+    def is_graph_valid(G):
+        """Check if model_graph is valid to build a executable
+
+        ``mmodel`` does not allow cycle graph, graph with isolated nodes,
+        and all nodes have callable attributes defined.
+        The method is bind to Model class because the following features
+        are specific for ``Model`` class
+        """
+
+        assert nx.is_directed(G), "invalid graph: undirected graph"
+        assert not nx.recursive_simple_cycles(G), "invalid graph: graph contains cycles"
+        assert not list(nx.isolates(G)), "invalid graph: graph contains isolated nodes"
+
+        assert is_node_attr_defined(
+            G, "obj"
+        ), "invalid graph: graph contains nodes with undefined callables"
+
+        # the following might occur when the node object is incorrectly constructed
+        assert is_node_attr_defined(G, "returns"), (
+            "invalid graph: graph contains nodes with undefined callables returns, "
+            "recommend using ModelGraph add_node_object method to add node object"
+        )
+        assert is_node_attr_defined(G, "sig"), (
+            "invalid graph: graph contains nodes with undefined callables signatures, "
+            "recommend using ModelGraph add_node_object method to add node object"
+        )
+        assert is_edge_attr_defined(G, 'val'), (
+            "invalid graph: graph contains edges with undefined variable attributes, "
+            "recommend using ModelGraph add_node_object method to add node object"
+        )
+
+        return True